Fixed array indexing by integer which was preventing JC kit converter from working correctly.

+++ src/com/redhat/ckey/applet/CardEdge.java
@@ -2244,19 +2244,19 @@ public class CardEdge extends Applet
 	short wrappedKeyFieldOffset = (short)(ISO7816.OFFSET_CDATA + 4);
 
 	// AC: Check the key type is as we expect
-	byte macKeyType = buffer[wrappedKeyFieldOffset + 0];
+	byte macKeyType = buffer[(short)(wrappedKeyFieldOffset + 0)];
 	if (macKeyType != 0x80){
 		ISOException.throwIt(SW_KEY_TYPE_INVALID);
 	}
 	
 	// AC: Check that the key size is valid (i.e. enough data exists in the buffer)
-	short macKeySize = (short)(buffer[wrappedKeyFieldOffset + 1] & 0x00FF);
+	short macKeySize = (short)(buffer[(short)(wrappedKeyFieldOffset + 1)] & 0x00FF);
 	if (macKeySize != 0x10){
 		ISOException.throwIt(SW_KEY_SIZE_ERROR);
 	}
 	
 	// AC: Check that key check size is valid (must be zero due to silly length == 23 check above)
-	short macKeyCheckSize = (short)(buffer[wrappedKeyFieldOffset + 2 + 0x10] & 0x00FF);
+	short macKeyCheckSize = (short)(buffer[(short)(wrappedKeyFieldOffset + 2 + 0x10)] & 0x00FF);
 	if (macKeyCheckSize != 0){ 
 		ISOException.throwIt(SW_UNSUPPORTED_FEATURE);
 	}
